requirements.txt
win10toast
requests
python-dotenv
pyperclip
keyboard